refactor: 대소변 시간 포맷 메서드(오전, 오후 ver) utils로 이동
시간 포맷 메서드(오전, 오후 ver) ExcretaAdapter에서 ExcretaDateTimeUtils로 이동
